phone-test/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
aiomysql==0.2.0
aiohttp==3.10.0
pyyaml==6.0.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
alembic==1.13.0
pydantic-settings==2.5.0
websockets==12.0
Pillow==10.4.0
numpy>=1.24.0
opencv-python-headless>=4.9.0
paddlepaddle>=2.6.0
paddleocr>=2.7.0
scikit-image>=0.21.0
aiosqlite>=0.20.0
